useMapに分割
